--- a/src/buffalo/index.ts
+++ b/src/buffalo/index.ts
@@ -1,4 +1,3 @@
 import Buffalo from './buffalo';
-import * as TsType from './tstype';
 
-export {Buffalo, TsType};
+export {Buffalo};