--- a/src/zcl/tstype.ts
+++ b/src/zcl/tstype.ts
@@ -16,7 +16,7 @@ interface Parameter {
 interface Command {
     ID: number;
     name: string;
-    parameters: Parameter[];
+    parameters: readonly Parameter[];
     response?: number;
 }