- Uppercased Zcl `DataType` (long overdue) and added some comments for good measure
- Isolated `Buffalo` and `BuffaloZcl` so they no longer have traces of zstack-specific logic.
  - `Buffalo` deals with base types and holds the generic functions.
  - `BuffaloZcl` adds support for Zcl-specific types.
  - Stack-specific buffalo can add support for stack-specific types as needed.
- Use enums directly instead of string comparisons for various Buffalo classes.
- Added missing types `read`/`write` to `BuffaloZcl` with tests.
- Added tests for missing coverage previously hidden behind "generic" `read`/`write`.
- Added extra checks for `options`-dependent and "non-value" `read`/`write` and tests to go with it.
  - _Some `write` are still missing "non-value" support, impossible because of current implementations. Requires refactoring._
- Added loads of typing and cleaned up old ones.
- Added missing Foundation commands with `BuffaloZcl` support and tests.

TODO: 
- [x] @Koenkk If you can confirm that zstack is happy with the changes. 🤞
- [x] https://github.com/Koenkk/zigbee-herdsman-converters/pull/7456
- [ ] `LONG_OCTET_STR` was previously using the same `read`/`write` as `LONG_CHAR_STR`, changed to `number[]`/`Buffer`. Need to check on the 2 use cases:
  - `Clusters.liXeePrivate.attributes.daysProfileCurrentCalendar`
  - `Clusters.liXeePrivate.attributes.daysProfileNextCalendar`
- [ ] Fix inconsistencies with args/return values for some uint/int types (should be plain numbers instead of lsb/msb tuples) then move these basic type `read`/`write` from `BuffaloZcl` to `Buffalo`. _Should probably be in another PR, will affect more than the Zcl definition._

NOTE: zigate missing `ParameterType` values were added arbitrarily to allow compiling. It seems to also be missing types support. Code looks broken overall.